debug chatmanager memory access issue
